delete-question-use-case

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -1,18 +1,24 @@
 import { QuestionsRepository } from "../repositories/questions-repository";
 
 interface deleteQuestionData {
-  id: string;
+    question_id: string;
+    author_id: string;
 }
 
 export class DeleteQuestion {
-  constructor(private questionsRepository: QuestionsRepository) {}
+    constructor(private questionsRepository: QuestionsRepository) {}
 
-  async execute(data: deleteQuestionData) {
-    const question = await this.questionsRepository.findById(data.id);
-    if (!question) {
-      throw new Error("Question not found");
+    async execute(data: deleteQuestionData) {
+        const question = await this.questionsRepository.findById(
+            data.question_id
+        );
+        if (!question) {
+            throw new Error("Question not found");
+        }
+        if (question.authorId.toString() !== data.author_id) {
+            throw new Error("You are not allowed to delete this question");
+        }
+        await this.questionsRepository.delete(question);
+        return;
     }
-    await this.questionsRepository.delete(question);
-    return;
-  }
 }
